+++ src/main/java/com/eoin/som/dao/StudentRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import javax.transaction.Transactional;
 
-public interface StudentRepository extends JpaRepository<Student, Integer> {
+public interface StudentRepository extends JpaRepository<Student, Long> {
     @Transactional
-    int deleteStudentById(int id);
+    Long deleteStudentById(Long id);
 }
