+++ src/main/java/com/eoin/som/dao/TeacherRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import javax.transaction.Transactional;
 
-public interface TeacherRepository extends JpaRepository<Teacher, Integer> {
+public interface TeacherRepository extends JpaRepository<Teacher, Long> {
     @Transactional
-    int deleteTeacherById(int id);
+    Long deleteTeacherById(Long id);
 }
